multiple cpp classes
