+++ fenics_mpm/cpp/MPMModel.h
@@ -4,6 +4,7 @@
 #include <dolfin/function/FunctionSpace.h>
 #include <dolfin/geometry/BoundingBoxTree.h>
 #include <dolfin/fem/GenericDofMap.h>
+#include "MPMMaterial.h"
 
 namespace dolfin
 {
@@ -15,6 +16,7 @@ namespace dolfin
       std::vector<double>       get_phi()      {return phi;};
       std::vector<unsigned int> get_vrt()      {return vrt;};
       std::vector<double>       get_grad_phi() {return grad_phi;};
+      void add_material(const MPMMaterial& M);
 
     private:
       const unsigned int cell_orientation = 0;
@@ -30,6 +32,7 @@ namespace dolfin
       std::size_t sdim;
       std::shared_ptr<const dolfin::Mesh> mesh;
       std::vector<double> vertex_coordinates;
+      std::vector<const MPMMaterial*> materials;
   };
 }
 #endif
