+++ fenics_mpm/gridmodel.py
@@ -3,9 +3,6 @@
 from   fenics            import *
 from   fenics_mpm.helper import print_text, print_min_max
 import numpy                 as np
-import os
-import inspect
-import instant
 
 
 class GridModel(object):
@@ -25,70 +22,6 @@ class GridModel(object):
     self.this = self
     #self.this = super(type(self), self)  # pointer to this base class
 
-    # open the cpp code for evaluating the basis functions :
-    cpp_src_dir = os.path.dirname(os.path.abspath(__file__)) + "/cpp/"
-    header_file = open(cpp_src_dir + "MPMModel.h", "r")
-    code        = header_file.read()
-    header_file.close()
-
-    system_headers = ['numpy/arrayobject.h',
-                      'dolfin/geometry/BoundingBoxTree.h',
-                      'dolfin/fem/GenericDofMap.h',
-                      'dolfin/function/FunctionSpace.h']
-    swigargs       = ['-c++', '-fcompact', '-O', '-I.', '-small']
-    cmake_packages = ['DOLFIN']
-    module_name    = "MPMModel"
-    sources        = ["MPMModel.cpp"]
-    source_dir     = cpp_src_dir
-    include_dirs   = [".", cpp_src_dir, 
-                      '/usr/lib/petscdir/3.7.3/x86_64-linux-gnu-real/include/']
-    module_name    = "MPMModel"
-    additional_decl = """
-    %init%{
-      import_array();
-      %}
-
-      // Include global SWIG interface files:
-      // Typemaps, shared_ptr declarations, exceptions, version
-      %include <boost_shared_ptr.i>
-
-      // Global typemaps and forward declarations
-      %include "dolfin/swig/typemaps/includes.i"
-      %include "dolfin/swig/forwarddeclarations.i"
-
-      // Global exceptions
-      %include <exception.i>
-
-      // Local shared_ptr declarations
-      %shared_ptr(dolfin::Function)
-      %shared_ptr(dolfin::FunctionSpace)
-
-      // %import types from submodule function of SWIG module function
-      %import(module="dolfin.cpp.function") "dolfin/function/Function.h"
-      %import(module="dolfin.cpp.function") "dolfin/function/FunctionSpace.h"
-
-      %feature("autodoc", "1");
-    """
-    #compiled_module = instant.build_module(
-    #    modulename = module_name,
-    #    code=code,
-    #    source_directory=source_dir,
-    #    additional_declarations=additional_decl,
-    #    system_headers=system_headers,
-    #    include_dirs=include_dirs,
-    #    swigargs=swigargs,
-    #    sources=sources,
-    #    cmake_packages=cmake_packages)
-
-    # compile this with Instant JIT compiler :
-    inst_params = {'code'                      : code,
-                   'module_name'               : module_name,
-                   'source_directory'          : cpp_src_dir,
-                   'sources'                   : sources,
-                   'additional_system_headers' : [],
-                   'include_dirs'              : include_dirs}
-    self.mpm_module = compile_extension_module(**inst_params)
-    
     # have the compiler generate code for evaluating basis derivatives :
     parameters['form_compiler']['no-evaluate_basis_derivatives'] = False
   
@@ -143,9 +76,6 @@ class GridModel(object):
     self.Q       = FunctionSpace(self.mesh, space, order)
     self.V       = VectorFunctionSpace(self.mesh, space, order)
     self.T       = TensorFunctionSpace(self.mesh, space, order)
-
-    # create a Probe instance from mpm_model.cpp :
-    self.probe = self.mpm_module.MPMModel(self.Q)
     
     s = "    - fundamental function spaces created - "
     print_text(s, cls=self.this)
@@ -238,71 +168,6 @@ class GridModel(object):
     self.assf_int_x = FunctionAssigner(self.f_int_x.function_space(), self.Q)
     self.assf_int_y = FunctionAssigner(self.f_int_y.function_space(), self.Q)
     self.assm       = FunctionAssigner(self.m.function_space(),       self.Q)
-
-  def get_particle_basis_functions(self, x):
-    r"""
-    Create particle basis functions for the single particle coordinate vector :math:`\mathbf{x}_p`.  In what follows, the topological dimension of the mesh is :math:`d` and the number of nodes per element is :math:`n_n`.
-
-    Returns a :py:obj:`tuple` of :class:`~numpy.ndarray`\s, respectively :
-
-    * ``vrt`` -- set of :math:`n_n` node indices :math:`i` for this particle's cell.  Shape is :math:`1 \times n_n`.
-    * ``phi`` -- interpolation function values :math:`\phi_i(\mathbf{x}_p)` corresponding to each node :math:`i` in ``vrt``.  Shape is :math:`1 \times n_n`.
-    * ``grad_phi`` -- gradient of interpolation function values :math:`\nabla \phi_i(\mathbf{x}_p)` corresponding to each node :math:`i` in ``vrt``.  Shape is :math:`d \times n_n`.  For example, in three dimensions :math:`\mathbf{x}_p = [x\ y\ z]^{\intercal}`, and with nodal indicies :math:`1,2,\ldots,n` ``grad_phi`` is
-
-    .. math:: \nabla \phi_i(\mathbf{x}_p) = 
-                                   \begin{bmatrix}
-                                      \frac{\partial \phi_1}{\partial x} & \frac{\partial \phi_1}{\partial y} & \frac{\partial \phi_1}{\partial z} \\
-                                      \frac{\partial \phi_2}{\partial x} & \frac{\partial \phi_2}{\partial y} & \frac{\partial \phi_2}{\partial z} \\
-                                      \vdots & \vdots & \vdots \\
-                                      \frac{\partial \phi_n}{\partial x} & \frac{\partial \phi_n}{\partial y} & \frac{\partial \phi_3}{\partial z}
-                                    \end{bmatrix}
-
-    :param x: global coordinate to evaluate.
-    :type x:  :py:obj:`int` or :py:obj:`float` or :class:`~numpy.ndarray`
-
-    :rtype: :py:obj:`tuple` (:class:`~numpy.ndarray`, :class:`~numpy.ndarray`, :class:`~numpy.ndarray`)
-    """
-    #mesh    = self.mesh
-    #element = self.element
-
-    ## find the cell with point :
-    #x_pt       = Point(x)
-    #cell_id    = mesh.bounding_box_tree().compute_first_entity_collision(x_pt)
-    #cell       = Cell(mesh, cell_id)
-    #coord_dofs = cell.get_vertex_coordinates()       # local coordinates
-    #
-    ## array for all basis functions of the cell :
-    #phi = np.zeros(element.space_dimension(), dtype=float)
-    #
-    ## array for values with derivatives of all 
-    ## basis functions, 2 * element dim :
-    #grad_phi = np.zeros(2*element.space_dimension(), dtype=float)
-    #
-    ## compute basis function values :
-    #element.evaluate_basis_all(phi, x, coord_dofs, cell.orientation())
-    #
-    ## compute 1st order derivatives :
-    #element.evaluate_basis_derivatives_all(1, grad_phi, x, 
-    #                                       coord_dofs, cell.orientation())
-
-    ## reshape such that rows are [d/dx, d/dy] :
-    #grad_phi = grad_phi.reshape((-1, 2))
-
-    ## get corresponding vertex indices, in dof indicies : 
-    #vrt = self.dofmap.cell_dofs(cell.index())
-    self.probe.eval(x);
-    vrt      = self.probe.get_vrt()
-    phi      = self.probe.get_phi()
-    grad_phi = self.probe.get_grad_phi()
-
-    # reshape such that rows are [d/dx, d/dy] :
-    grad_phi = grad_phi.reshape((-1, 2))
-
-    #print "vrt =", type(vrt), vrt
-    #print "phi =", type(phi), phi
-    #print "grad_phi = ", type(grad_phi), grad_phi
-
-    return vrt, phi, grad_phi
 
   def update_mass(self, m):
     r"""
